Always emit logs from exec
Today, lock-exec doesn't appear to output any logs of the child process
if it errors out. `CombinedOutput` will always return the buffer that
contains the stderr/stdout. By running the print logic before
returning from the error, we can still get the logs of the child
process regardless of how it exited.