Allow failures due to lock contention fail successfully
This will enable tracking actual errors like failed commands or other
misc dynamodb errors. Lock contention is an expected behavior and we
shouldn't be treating those as errors.

I also updated the `Run` function to always fallback to returning the
error from the command execution and not exiting early.

I also stuck in a strings.TrimSpace.